Exp: Adding generator_1 as an example.
Trying to refactor spaced to offer infinite schedules while limiting the
amount of memory required by each object.  The generator_1.py file is an
example of basic features required by repetition.py.  If more are
required they will be added, but for now it is an good example to work
from.